cli/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "skillhub"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "requests>=2.28",
    "PyYAML>=6.0",
    "pathspec>=0.12",
    "packaging>=24.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-mock>=3.0"]

[project.scripts]
skillhub = "skillhub.cli:cli"
